Chnages to improve the security of the code.
